chore(federation): move operation code to a top-level module (#5322)
