+++ apollo-federation/src/query_plan/conditions.rs
@@ -9,10 +9,10 @@ use indexmap::map::Entry;
 use indexmap::IndexMap;
 
 use crate::error::FederationError;
+use crate::operation::Selection;
+use crate::operation::SelectionMap;
+use crate::operation::SelectionSet;
 use crate::query_graph::graph_path::OpPathElement;
-use crate::query_plan::operation::Selection;
-use crate::query_plan::operation::SelectionMap;
-use crate::query_plan::operation::SelectionSet;
 
 /// This struct is meant for tracking whether a selection set in a `FetchDependencyGraphNode` needs
 /// to be queried, based on the `@skip`/`@include` applications on the selections within.
