+++ apollo-federation/src/schema/field_set.rs
@@ -10,8 +10,8 @@ use indexmap::IndexMap;
 use crate::error::FederationError;
 use crate::error::MultipleFederationErrors;
 use crate::error::SingleFederationError;
-use crate::query_plan::operation::NamedFragments;
-use crate::query_plan::operation::SelectionSet;
+use crate::operation::NamedFragments;
+use crate::operation::SelectionSet;
 use crate::schema::position::CompositeTypeDefinitionPosition;
 use crate::schema::position::FieldDefinitionPosition;
 use crate::schema::position::InterfaceTypeDefinitionPosition;
