+++ apollo-federation/src/schema/subgraph_metadata.rs
@@ -6,8 +6,8 @@ use crate::error::FederationError;
 use crate::link::federation_spec_definition::FederationSpecDefinition;
 use crate::link::spec::Version;
 use crate::link::spec_definition::SpecDefinition;
-use crate::query_plan::operation::Selection;
-use crate::query_plan::operation::SelectionSet;
+use crate::operation::Selection;
+use crate::operation::SelectionSet;
 use crate::schema::field_set::add_interface_field_implementations;
 use crate::schema::field_set::collect_target_fields_from_field_set;
 use crate::schema::position::CompositeTypeDefinitionPosition;
